Upgrade code design
